IGNITE-5737 Fix broken identity revert endpoint.
Error middleware which was added earlier this branch handles missing response value as errors, which in turn was causing 404 instead 200 for identity revert requests.
I've fixed the issue by adding a placeholder return value to sessions.revert service method. The actual value isn't used anywhere, so it does not matter.